% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{AdministrationRegion}
\alias{AdministrationRegion}
\title{Google Civic Information API Objects 
Provides polling places, early vote locations, contest data, election officials, and government representatives for U.S. residential addresses.}
\usage{
AdministrationRegion(electionAdministrationBody = NULL, id = NULL,
  local_jurisdiction = NULL, name = NULL, sources = NULL)
}
\arguments{
\item{electionAdministrationBody}{The election administration body for this area}

\item{id}{An ID for this object}

\item{local_jurisdiction}{The city or county that provides election information for this voter}

\item{name}{The name of the jurisdiction}

\item{sources}{A list of sources for this area}
}
\value{
AdministrationRegion object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:31:01
filename: /Users/mark/dev/R/autoGoogleAPI/googlecivicinfov2.auto/R/civicinfo_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AdministrationRegion Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes information about a regional election administrative area.
}
