% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{AdministrativeBody}
\alias{AdministrativeBody}
\title{AdministrativeBody Object}
\usage{
AdministrativeBody(absenteeVotingInfoUrl = NULL, addressLines = NULL,
  ballotInfoUrl = NULL, correspondenceAddress = NULL,
  electionInfoUrl = NULL, electionOfficials = NULL,
  electionRegistrationConfirmationUrl = NULL,
  electionRegistrationUrl = NULL, electionRulesUrl = NULL,
  hoursOfOperation = NULL, name = NULL, physicalAddress = NULL,
  voter_services = NULL, votingLocationFinderUrl = NULL)
}
\arguments{
\item{absenteeVotingInfoUrl}{A URL provided by this administrative body for information on absentee voting}

\item{addressLines}{No description}

\item{ballotInfoUrl}{A URL provided by this administrative body to give contest information to the voter}

\item{correspondenceAddress}{The mailing address of this administrative body}

\item{electionInfoUrl}{A URL provided by this administrative body for looking up general election information}

\item{electionOfficials}{The election officials for this election administrative body}

\item{electionRegistrationConfirmationUrl}{A URL provided by this administrative body for confirming that the voter is registered to vote}

\item{electionRegistrationUrl}{A URL provided by this administrative body for looking up how to register to vote}

\item{electionRulesUrl}{A URL provided by this administrative body describing election rules to the voter}

\item{hoursOfOperation}{A description of the hours of operation for this administrative body}

\item{name}{The name of this election administrative body}

\item{physicalAddress}{The physical address of this administrative body}

\item{voter_services}{A description of the services this administrative body may provide}

\item{votingLocationFinderUrl}{A URL provided by this administrative body for looking up where to vote}
}
\value{
AdministrativeBody object
}
\description{
AdministrativeBody Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about an election administrative body (e.g. County Board of Elections).
}
