% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{Candidate}
\alias{Candidate}
\title{Candidate Object}
\usage{
Candidate(candidateUrl = NULL, channels = NULL, email = NULL,
  name = NULL, orderOnBallot = NULL, party = NULL, phone = NULL,
  photoUrl = NULL)
}
\arguments{
\item{candidateUrl}{The URL for the candidate's campaign web site}

\item{channels}{A list of known (social) media channels for this candidate}

\item{email}{The email address for the candidate's campaign}

\item{name}{The candidate's name}

\item{orderOnBallot}{The order the candidate appears on the ballot for this contest}

\item{party}{The full name of the party the candidate is a member of}

\item{phone}{The voice phone number for the candidate's campaign office}

\item{photoUrl}{A URL for a photo of the candidate}
}
\value{
Candidate object
}
\description{
Candidate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a candidate running for elected office.
}
