% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{Channel}
\alias{Channel}
\title{Channel Object}
\usage{
Channel(id = NULL, type = NULL)
}
\arguments{
\item{id}{The unique public identifier for the candidate's channel}

\item{type}{The type of channel}
}
\value{
Channel object
}
\description{
Channel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A social media or web channel for a candidate.
}
