% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_functions.R
\docType{package}
\name{civicinfo_googleAuthR}
\alias{civicinfo_googleAuthR}
\alias{civicinfo_googleAuthR-package}
\title{Google Civic Information API
Provides polling places, early vote locations, contest data, election officials, and government representatives for U.S. residential addresses.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:31:00
filename: /Users/mark/dev/R/autoGoogleAPI/googlecivicinfov2.auto/R/civicinfo_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}
