% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{Contest}
\alias{Contest}
\title{Contest Object}
\usage{
Contest(ballotPlacement = NULL, candidates = NULL, district = NULL,
  electorateSpecifications = NULL, id = NULL, level = NULL,
  numberElected = NULL, numberVotingFor = NULL, office = NULL,
  primaryParty = NULL, referendumBallotResponses = NULL,
  referendumBrief = NULL, referendumConStatement = NULL,
  referendumEffectOfAbstain = NULL, referendumPassageThreshold = NULL,
  referendumProStatement = NULL, referendumSubtitle = NULL,
  referendumText = NULL, referendumTitle = NULL, referendumUrl = NULL,
  roles = NULL, sources = NULL, special = NULL, type = NULL)
}
\arguments{
\item{ballotPlacement}{A number specifying the position of this contest on the voter's ballot}

\item{candidates}{The candidate choices for this contest}

\item{district}{Information about the electoral district that this contest is in}

\item{electorateSpecifications}{A description of any additional eligibility requirements for voting in this contest}

\item{id}{An ID for this object}

\item{level}{The levels of government of the office for this contest}

\item{numberElected}{The number of candidates that will be elected to office in this contest}

\item{numberVotingFor}{The number of candidates that a voter may vote for in this contest}

\item{office}{The name of the office for this contest}

\item{primaryParty}{If this is a partisan election, the name of the party it is for}

\item{referendumBallotResponses}{The set of ballot responses for the referendum}

\item{referendumBrief}{Specifies a short summary of the referendum that is typically on the ballot below the title but above the text}

\item{referendumConStatement}{A statement in opposition to the referendum}

\item{referendumEffectOfAbstain}{Specifies what effect abstaining (not voting) on the proposition will have (i}

\item{referendumPassageThreshold}{The threshold of votes that the referendum needs in order to pass, e}

\item{referendumProStatement}{A statement in favor of the referendum}

\item{referendumSubtitle}{A brief description of the referendum}

\item{referendumText}{The full text of the referendum}

\item{referendumTitle}{The title of the referendum (e}

\item{referendumUrl}{A link to the referendum}

\item{roles}{The roles which this office fulfills}

\item{sources}{A list of sources for this contest}

\item{special}{'Yes' or 'No' depending on whether this a contest being held outside the normal election cycle}

\item{type}{The type of contest}
}
\value{
Contest object
}
\description{
Contest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a contest that appears on a voter's ballot.
}

