% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{ContextParams}
\alias{ContextParams}
\title{ContextParams Object}
\usage{
ContextParams(clientProfile = NULL)
}
\arguments{
\item{clientProfile}{No description}
}
\value{
ContextParams object
}
\description{
ContextParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
