% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{DivisionRepresentativeInfoRequest}
\alias{DivisionRepresentativeInfoRequest}
\title{DivisionRepresentativeInfoRequest Object}
\usage{
DivisionRepresentativeInfoRequest(contextParams = NULL)
}
\arguments{
\item{contextParams}{No description}
}
\value{
DivisionRepresentativeInfoRequest object
}
\description{
DivisionRepresentativeInfoRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to look up representative information for a single division.
}
\seealso{
Other DivisionRepresentativeInfoRequest functions: \code{\link{representatives.representativeInfoByDivision}}
}
