% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_functions.R
\name{divisions.search}
\alias{divisions.search}
\title{Searches for political divisions by their natural name or OCD ID.}
\usage{
divisions.search(DivisionSearchRequest, query = NULL)
}
\arguments{
\item{DivisionSearchRequest}{The \link{DivisionSearchRequest} object to pass to this method}

\item{query}{The search query}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/civic-information}{Google Documentation}

Other DivisionSearchRequest functions: \code{\link{DivisionSearchRequest}}
}
