% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{DivisionSearchRequest}
\alias{DivisionSearchRequest}
\title{DivisionSearchRequest Object}
\usage{
DivisionSearchRequest(contextParams = NULL)
}
\arguments{
\item{contextParams}{No description}
}
\value{
DivisionSearchRequest object
}
\description{
DivisionSearchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A search request for political geographies.
}
\seealso{
Other DivisionSearchRequest functions: \code{\link{divisions.search}}
}
