% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{DivisionSearchResponse}
\alias{DivisionSearchResponse}
\title{DivisionSearchResponse Object}
\usage{
DivisionSearchResponse(results = NULL)
}
\arguments{
\item{results}{No description}
}
\value{
DivisionSearchResponse object
}
\description{
DivisionSearchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of a division search query.
}
