% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{DivisionSearchResult}
\alias{DivisionSearchResult}
\title{DivisionSearchResult Object}
\usage{
DivisionSearchResult(aliases = NULL, name = NULL, ocdId = NULL)
}
\arguments{
\item{aliases}{Other Open Civic Data identifiers that refer to the same division -- for example, those that refer to other political divisions whose boundaries are defined to be coterminous with this one}

\item{name}{The name of the division}

\item{ocdId}{The unique Open Civic Data identifier for this division}
}
\value{
DivisionSearchResult object
}
\description{
DivisionSearchResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a political geographic division that matches the requested query.
}

