% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{Election}
\alias{Election}
\title{Election Object}
\usage{
Election(electionDay = NULL, id = NULL, name = NULL,
  ocdDivisionId = NULL)
}
\arguments{
\item{electionDay}{Day of the election in YYYY-MM-DD format}

\item{id}{The unique ID of this election}

\item{name}{A displayable name for the election}

\item{ocdDivisionId}{The political division of the election}
}
\value{
Election object
}
\description{
Election Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about the election that was queried.
}

