% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{ElectionOfficial}
\alias{ElectionOfficial}
\title{ElectionOfficial Object}
\usage{
ElectionOfficial(emailAddress = NULL, faxNumber = NULL, name = NULL,
  officePhoneNumber = NULL, title = NULL)
}
\arguments{
\item{emailAddress}{The email address of the election official}

\item{faxNumber}{The fax number of the election official}

\item{name}{The full name of the election official}

\item{officePhoneNumber}{The office phone number of the election official}

\item{title}{The title of the election official}
}
\value{
ElectionOfficial object
}
\description{
ElectionOfficial Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about individual election officials.
}

