% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_functions.R
\name{elections.electionQuery}
\alias{elections.electionQuery}
\title{List of available elections to query.}
\usage{
elections.electionQuery(ElectionsQueryRequest)
}
\arguments{
\item{ElectionsQueryRequest}{The \link{ElectionsQueryRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/civic-information}{Google Documentation}

Other ElectionsQueryRequest functions: \code{\link{ElectionsQueryRequest}}
}
