% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_functions.R
\name{elections.voterInfoQuery}
\alias{elections.voterInfoQuery}
\title{Looks up information relevant to a voter based on the voter's registered address.}
\usage{
elections.voterInfoQuery(VoterInfoRequest, address, electionId = NULL,
  officialOnly = NULL, returnAllAvailableData = NULL)
}
\arguments{
\item{VoterInfoRequest}{The \link{VoterInfoRequest} object to pass to this method}

\item{address}{The registered address of the voter to look up}

\item{electionId}{The unique ID of the election to look up}

\item{officialOnly}{If set to true, only data from official state sources will be returned}

\item{returnAllAvailableData}{If set to true, the query will return the success codeand include any partial information when it is unable to determine a matching address or unable to determine the election for electionId=0 queries}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/civic-information}{Google Documentation}

Other VoterInfoRequest functions: \code{\link{VoterInfoRequest}}
}
