% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{ElectionsQueryRequest}
\alias{ElectionsQueryRequest}
\title{ElectionsQueryRequest Object}
\usage{
ElectionsQueryRequest(contextParams = NULL)
}
\arguments{
\item{contextParams}{No description}
}
\value{
ElectionsQueryRequest object
}
\description{
ElectionsQueryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ElectionsQueryRequest functions: \code{\link{elections.electionQuery}}
}
