% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{ElectionsQueryResponse}
\alias{ElectionsQueryResponse}
\title{ElectionsQueryResponse Object}
\usage{
ElectionsQueryResponse(elections = NULL)
}
\arguments{
\item{elections}{A list of available elections}
}
\value{
ElectionsQueryResponse object
}
\description{
ElectionsQueryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The list of elections available for this version of the API.
}
