% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{ElectoralDistrict}
\alias{ElectoralDistrict}
\title{ElectoralDistrict Object}
\usage{
ElectoralDistrict(id = NULL, kgForeignKey = NULL, name = NULL,
  scope = NULL)
}
\arguments{
\item{id}{An identifier for this district, relative to its scope}

\item{kgForeignKey}{No description}

\item{name}{The name of the district}

\item{scope}{The geographic scope of this district}
}
\value{
ElectoralDistrict object
}
\description{
ElectoralDistrict Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the geographic scope of a contest.
}

