% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{GeographicDivision}
\alias{GeographicDivision}
\title{GeographicDivision Object}
\usage{
GeographicDivision(alsoKnownAs = NULL, name = NULL, officeIndices = NULL)
}
\arguments{
\item{alsoKnownAs}{Any other valid OCD IDs that refer to the same division}

\item{name}{The name of the division}

\item{officeIndices}{List of indices in the offices array, one for each office elected from this division}
}
\value{
GeographicDivision object
}
\description{
GeographicDivision Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a political geography.
}

