% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{Office}
\alias{Office}
\title{Office Object}
\usage{
Office(divisionId = NULL, levels = NULL, name = NULL,
  officialIndices = NULL, roles = NULL, sources = NULL)
}
\arguments{
\item{divisionId}{The OCD ID of the division with which this office is associated}

\item{levels}{The levels of government of which this office is part}

\item{name}{The human-readable name of the office}

\item{officialIndices}{List of indices in the officials array of people who presently hold this office}

\item{roles}{The roles which this office fulfills}

\item{sources}{A list of sources for this office}
}
\value{
Office object
}
\description{
Office Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about an Office held by one or more Officials.
}

