% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{Official}
\alias{Official}
\title{Official Object}
\usage{
Official(address = NULL, channels = NULL, emails = NULL, name = NULL,
  party = NULL, phones = NULL, photoUrl = NULL, urls = NULL)
}
\arguments{
\item{address}{Addresses at which to contact the official}

\item{channels}{A list of known (social) media channels for this official}

\item{emails}{The direct email addresses for the official}

\item{name}{The official's name}

\item{party}{The full name of the party the official belongs to}

\item{phones}{The official's public contact phone numbers}

\item{photoUrl}{A URL for a photo of the official}

\item{urls}{The official's public website URLs}
}
\value{
Official object
}
\description{
Official Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a person holding an elected office.
}

