% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{PollingLocation}
\alias{PollingLocation}
\title{PollingLocation Object}
\usage{
PollingLocation(address = NULL, endDate = NULL, id = NULL, name = NULL,
  notes = NULL, pollingHours = NULL, sources = NULL, startDate = NULL,
  voterServices = NULL)
}
\arguments{
\item{address}{The address of the location}

\item{endDate}{The last date that this early vote site or drop off location may be used}

\item{id}{An ID for this object}

\item{name}{The name of the early vote site or drop off location}

\item{notes}{Notes about this location (e}

\item{pollingHours}{A description of when this location is open}

\item{sources}{A list of sources for this location}

\item{startDate}{The first date that this early vote site or drop off location may be used}

\item{voterServices}{The services provided by this early vote site or drop off location}
}
\value{
PollingLocation object
}
\description{
PollingLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A location where a voter can vote. This may be an early vote site, an election day voting location, or a drop off location for a completed ballot.
}

