% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{PostalAddress}
\alias{PostalAddress}
\title{PostalAddress Object}
\usage{
PostalAddress(addressLines = NULL, administrativeAreaName = NULL,
  countryName = NULL, countryNameCode = NULL,
  dependentLocalityName = NULL, dependentThoroughfareLeadingType = NULL,
  dependentThoroughfareName = NULL,
  dependentThoroughfarePostDirection = NULL,
  dependentThoroughfarePreDirection = NULL,
  dependentThoroughfareTrailingType = NULL,
  dependentThoroughfaresConnector = NULL,
  dependentThoroughfaresIndicator = NULL, dependentThoroughfaresType = NULL,
  firmName = NULL, isDisputed = NULL, languageCode = NULL,
  localityName = NULL, postBoxNumber = NULL, postalCodeNumber = NULL,
  postalCodeNumberExtension = NULL, premiseName = NULL,
  recipientName = NULL, sortingCode = NULL,
  subAdministrativeAreaName = NULL, subPremiseName = NULL,
  thoroughfareLeadingType = NULL, thoroughfareName = NULL,
  thoroughfareNumber = NULL, thoroughfarePostDirection = NULL,
  thoroughfarePreDirection = NULL, thoroughfareTrailingType = NULL)
}
\arguments{
\item{addressLines}{No description}

\item{administrativeAreaName}{No description}

\item{countryName}{No description}

\item{countryNameCode}{No description}

\item{dependentLocalityName}{No description}

\item{dependentThoroughfareLeadingType}{No description}

\item{dependentThoroughfareName}{No description}

\item{dependentThoroughfarePostDirection}{No description}

\item{dependentThoroughfarePreDirection}{No description}

\item{dependentThoroughfareTrailingType}{No description}

\item{dependentThoroughfaresConnector}{No description}

\item{dependentThoroughfaresIndicator}{No description}

\item{dependentThoroughfaresType}{No description}

\item{firmName}{No description}

\item{isDisputed}{No description}

\item{languageCode}{No description}

\item{localityName}{No description}

\item{postBoxNumber}{No description}

\item{postalCodeNumber}{No description}

\item{postalCodeNumberExtension}{No description}

\item{premiseName}{No description}

\item{recipientName}{No description}

\item{sortingCode}{No description}

\item{subAdministrativeAreaName}{No description}

\item{subPremiseName}{No description}

\item{thoroughfareLeadingType}{No description}

\item{thoroughfareName}{No description}

\item{thoroughfareNumber}{No description}

\item{thoroughfarePostDirection}{No description}

\item{thoroughfarePreDirection}{No description}

\item{thoroughfareTrailingType}{No description}
}
\value{
PostalAddress object
}
\description{
PostalAddress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
