% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{RepresentativeInfoData.divisions}
\alias{RepresentativeInfoData.divisions}
\title{RepresentativeInfoData.divisions Object}
\usage{
RepresentativeInfoData.divisions()
}
\value{
RepresentativeInfoData.divisions object
}
\description{
RepresentativeInfoData.divisions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Political geographic divisions that contain the requested address.
}
\seealso{
Other RepresentativeInfoData functions: \code{\link{RepresentativeInfoData}}
}
