% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{RepresentativeInfoData}
\alias{RepresentativeInfoData}
\title{RepresentativeInfoData Object}
\usage{
RepresentativeInfoData(RepresentativeInfoData.divisions = NULL,
  divisions = NULL, offices = NULL, officials = NULL)
}
\arguments{
\item{RepresentativeInfoData.divisions}{The \link{RepresentativeInfoData.divisions} object or list of objects}

\item{divisions}{Political geographic divisions that contain the requested address}

\item{offices}{Elected offices referenced by the divisions listed above}

\item{officials}{Officials holding the offices listed above}
}
\value{
RepresentativeInfoData object
}
\description{
RepresentativeInfoData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RepresentativeInfoData functions: \code{\link{RepresentativeInfoData.divisions}}
}

