% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{RepresentativeInfoRequest}
\alias{RepresentativeInfoRequest}
\title{RepresentativeInfoRequest Object}
\usage{
RepresentativeInfoRequest(contextParams = NULL)
}
\arguments{
\item{contextParams}{No description}
}
\value{
RepresentativeInfoRequest object
}
\description{
RepresentativeInfoRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request for political geography and representative information for an address.
}
\seealso{
Other RepresentativeInfoRequest functions: \code{\link{representatives.representativeInfoByAddress}}
}
