% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{RepresentativeInfoResponse.divisions}
\alias{RepresentativeInfoResponse.divisions}
\title{RepresentativeInfoResponse.divisions Object}
\usage{
RepresentativeInfoResponse.divisions()
}
\value{
RepresentativeInfoResponse.divisions object
}
\description{
RepresentativeInfoResponse.divisions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Political geographic divisions that contain the requested address.
}
\seealso{
Other RepresentativeInfoResponse functions: \code{\link{RepresentativeInfoResponse}}
}

