% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{RepresentativeInfoResponse}
\alias{RepresentativeInfoResponse}
\title{RepresentativeInfoResponse Object}
\usage{
RepresentativeInfoResponse(RepresentativeInfoResponse.divisions = NULL,
  divisions = NULL, normalizedInput = NULL, offices = NULL,
  officials = NULL)
}
\arguments{
\item{RepresentativeInfoResponse.divisions}{The \link{RepresentativeInfoResponse.divisions} object or list of objects}

\item{divisions}{Political geographic divisions that contain the requested address}

\item{normalizedInput}{The normalized version of the requested address}

\item{offices}{Elected offices referenced by the divisions listed above}

\item{officials}{Officials holding the offices listed above}
}
\value{
RepresentativeInfoResponse object
}
\description{
RepresentativeInfoResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of a representative info lookup query.
}
\seealso{
Other RepresentativeInfoResponse functions: \code{\link{RepresentativeInfoResponse.divisions}}
}
