% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_functions.R
\name{representatives.representativeInfoByDivision}
\alias{representatives.representativeInfoByDivision}
\title{Looks up representative information for a single geographic division.}
\usage{
representatives.representativeInfoByDivision(DivisionRepresentativeInfoRequest,
  ocdId, levels = NULL, recursive = NULL, roles = NULL)
}
\arguments{
\item{DivisionRepresentativeInfoRequest}{The \link{DivisionRepresentativeInfoRequest} object to pass to this method}

\item{ocdId}{The Open Civic Data division identifier of the division to look up}

\item{levels}{A list of office levels to filter by}

\item{recursive}{If true, information about all divisions contained in the division requested will be included as well}

\item{roles}{A list of office roles to filter by}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/civic-information}{Google Documentation}

Other DivisionRepresentativeInfoRequest functions: \code{\link{DivisionRepresentativeInfoRequest}}
}
