% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{SimpleAddressType}
\alias{SimpleAddressType}
\title{SimpleAddressType Object}
\usage{
SimpleAddressType(city = NULL, line1 = NULL, line2 = NULL, line3 = NULL,
  locationName = NULL, state = NULL, zip = NULL)
}
\arguments{
\item{city}{The city or town for the address}

\item{line1}{The street name and number of this address}

\item{line2}{The second line the address, if needed}

\item{line3}{The third line of the address, if needed}

\item{locationName}{The name of the location}

\item{state}{The US two letter state abbreviation of the address}

\item{zip}{The US Postal Zip Code of the address}
}
\value{
SimpleAddressType object
}
\description{
SimpleAddressType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A simple representation of an address.
}

