% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{Source}
\alias{Source}
\title{Source Object}
\usage{
Source(name = NULL, official = NULL)
}
\arguments{
\item{name}{The name of the data source}

\item{official}{Whether this data comes from an official government source}
}
\value{
Source object
}
\description{
Source Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about the data source for the element containing it.
}
