% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{VoterInfoRequest}
\alias{VoterInfoRequest}
\title{VoterInfoRequest Object}
\usage{
VoterInfoRequest(contextParams = NULL, voterInfoSegmentResult = NULL)
}
\arguments{
\item{contextParams}{No description}

\item{voterInfoSegmentResult}{No description}
}
\value{
VoterInfoRequest object
}
\description{
VoterInfoRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request for information about a voter.
}
\seealso{
Other VoterInfoRequest functions: \code{\link{elections.voterInfoQuery}}
}
