% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{VoterInfoResponse}
\alias{VoterInfoResponse}
\title{VoterInfoResponse Object}
\usage{
VoterInfoResponse(contests = NULL, dropOffLocations = NULL,
  earlyVoteSites = NULL, election = NULL, mailOnly = NULL,
  normalizedInput = NULL, otherElections = NULL, pollingLocations = NULL,
  precinctId = NULL, state = NULL)
}
\arguments{
\item{contests}{Contests that will appear on the voter's ballot}

\item{dropOffLocations}{Locations where a voter is eligible to drop off a completed ballot}

\item{earlyVoteSites}{Locations where the voter is eligible to vote early, prior to election day}

\item{election}{The election that was queried}

\item{mailOnly}{Specifies whether voters in the precinct vote only by mailing their ballots (with the possible option of dropping off their ballots as well)}

\item{normalizedInput}{The normalized version of the requested address}

\item{otherElections}{If no election ID was specified in the query, and there was more than one election with data for the given voter, this will contain information about the other elections that could apply}

\item{pollingLocations}{Locations where the voter is eligible to vote on election day}

\item{precinctId}{No description}

\item{state}{Local Election Information for the state that the voter votes in}
}
\value{
VoterInfoResponse object
}
\description{
VoterInfoResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of a voter info lookup query.
}

