% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicinfo_objects.R
\name{VoterInfoSegmentResult}
\alias{VoterInfoSegmentResult}
\title{VoterInfoSegmentResult Object}
\usage{
VoterInfoSegmentResult(generatedMillis = NULL, postalAddress = NULL,
  request = NULL, response = NULL)
}
\arguments{
\item{generatedMillis}{No description}

\item{postalAddress}{No description}

\item{request}{No description}

\item{response}{No description}
}
\value{
VoterInfoSegmentResult object
}
\description{
VoterInfoSegmentResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
